Move AWS discovery code to command/agent/config_aws.go
